+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Adminer dev
 - Ignore invalid X-Forwarded-Prefix (GHSA-fr74-9mf9-gf44)
 - Compile: Valid UTF-8 (bug #1113, #1213)
+- Compile: Replace LZW with deflate compression
 - Add missing translations
 - Syntax highlighting: Update SQLite and PostgreSQL
 - Support disabled set_time_limit() and ini_set() (bug #1288)